chore(RHINENG-11759): Remove feature flag
The user acknowledgement of new matches feature has been hidden behind a feature flag in production up until this point. Now that it is ready to go live, we are removing the feature flag.
