+++ src/Components/SigDetailsTable/SigDetailsTable.js
@@ -1,45 +1,40 @@
-import React, { useEffect, useReducer } from 'react';
+import React, { useEffect, useReducer, useState } from 'react';
 import propTypes from 'prop-types';
 import { useIntl } from 'react-intl';
-import { useQuery } from '@apollo/client';
+import { useApolloClient } from '@apollo/client';
 import { isEmpty, uniqWith } from 'lodash';
 import { Pagination, PaginationVariant } from '@patternfly/react-core';
-import {
-  Table,
-  TableBody,
-  TableHeader,
-} from '@patternfly/react-table/deprecated';
-import {
-  SortByDirection,
-  cellWidth,
-  expandable,
-  sortable,
-  info,
-} from '@patternfly/react-table';
+import { Table } from '@patternfly/react-table';
+import { SortByDirection } from '@patternfly/react-table';
 import { CheckCircleIcon, SearchIcon } from '@patternfly/react-icons';
-import InsightsLink from '@redhat-cloud-services/frontend-components/InsightsLink';
-import SkeletonTable from '@patternfly/react-component-groups/dist/esm/SkeletonTable';
+import { SkeletonTable } from '@patternfly/react-component-groups';
 import { PrimaryToolbar } from '@redhat-cloud-services/frontend-components/PrimaryToolbar';
 import TableToolbar from '@redhat-cloud-services/frontend-components/TableToolbar';
-import { DateFormat } from '@redhat-cloud-services/frontend-components/DateFormat';
-import { expandMatchMetadata } from '../Common';
 import {
   GET_SIGNATURE_DETAILS_TABLE,
   GET_SIGNATURE_DETAILS_TABLE_GROUPS,
 } from '../../operations/queries';
-import CodeEditor from '../CodeEditor/CodeEditor';
 import MessageState from '../MessageState/MessageState';
 import messages from '../../Messages';
 import EmptyAccount from '../SharedComponents/EmptyAccount';
 import './SigDetailsTable.scss';
 import { useWorkspaceFeatureFlag } from '../../Utilities/Hooks';
+import { matchStatusFilterMap } from '../helpers';
+import { sigDetailsTableColumns } from './Columns';
+import { useActionsConfig } from '../../Utilities/Hooks';
+import { filterValues } from '../constants';
+import { usePermissions } from '@redhat-cloud-services/frontend-components-utilities/RBACHook';
+import { RBACPermissions } from '../Common';
+import { DetailsTableHeader } from '../TableComponents/DetailsTableHeader';
+import { SigDetailsTableBody } from './SigDetailsTableBody';
 
 const sortIndices = {
   1: 'DISPLAY_NAME',
   2: 'GROUP_NAME',
   3: 'OS_VERSION',
   4: 'LAST_SCAN_DATE',
-  5: 'MATCH_COUNT',
+  5: 'NEW_MATCH_COUNT',
+  6: 'MATCH_COUNT',
 };
 
 const orderBy = ({ index, direction }) =>
@@ -48,7 +43,19 @@ const orderBy = ({ index, direction }) =>
 const tableReducer = (state, action) => {
   switch (action.type) {
     case 'setTableVars':
-      return { ...state, tableVars: { ...state.tableVars, ...action.payload } };
+      return {
+        ...state,
+        tableVars:
+          action.payload.sigMatchStatusFilter?.length === 0
+            ? {
+                ...state.tableVars,
+                sigMatchStatusFilter: undefined,
+              }
+            : {
+                ...state.tableVars,
+                ...action.payload,
+              },
+      };
     case 'setSortBy':
       return {
         ...state,
@@ -66,6 +73,18 @@ const tableReducer = (state, action) => {
 
 const SigDetailsTable = ({ ruleName, affectedCount, isEmptyAccount }) => {
   const intl = useIntl();
+  const client = useApolloClient();
+  const isWorkspaceEnabled = useWorkspaceFeatureFlag();
+  const { hasAccess: hasWriteAccess, isLoading: isWriteAccessLoading } =
+    usePermissions('malware-detection', RBACPermissions.write);
+  const [selectedMatches, setSelectedMatches] = useState([]);
+  const [sigDetailsTableLoading, setSigDetailsTableLoading] = useState(false);
+  const [sigDetailsTable, setSigDetailsTable] = useState({});
+  const [sigDetailsTableGroupsLoading, setSigDetailsTableGroupsLoading] =
+    useState(false);
+  const [sigDetailsTableGroups, setSigDetailsTableGroups] = useState({});
+  const [expandedRows, setExpandedRows] = useState([]);
+
   const initialState = {
     tableVars: {
       limit: 10,
@@ -73,6 +92,7 @@ const SigDetailsTable = ({ ruleName, affectedCount, isEmptyAccount }) => {
       orderBy: 'LAST_SCAN_DATE_DESC',
       displayName: '',
       hostGroupFilter: undefined,
+      sigMatchStatusFilter: undefined,
       ruleName,
     },
     sortBy: {
@@ -82,53 +102,58 @@ const SigDetailsTable = ({ ruleName, affectedCount, isEmptyAccount }) => {
     rows: [],
     groups: [],
   };
-  const [{ tableVars, sortBy, rows, groups }, stateSet] = useReducer(
-    tableReducer,
-    {
-      ...initialState,
-    }
-  );
-  const { data, loading, error } = useQuery(GET_SIGNATURE_DETAILS_TABLE, {
-    variables: tableVars,
+
+  const [{ tableVars, sortBy, groups }, stateSet] = useReducer(tableReducer, {
+    ...initialState,
   });
-  const { data: groupsData, loading: groupsLoading } = useQuery(
-    GET_SIGNATURE_DETAILS_TABLE_GROUPS,
-    { variables: tableVars }
-  );
-  const columns = (isWorkspaceEnabled) => [
-    {
-      title: intl.formatMessage(messages.name),
-      cellFormatters: [expandable],
-      transforms: [sortable, cellWidth(45)],
-    },
-    {
-      title: isWorkspaceEnabled
-        ? 'Workspace'
-        : intl.formatMessage(messages.group),
-      transforms: [sortable, cellWidth(15)],
-    },
-    {
-      title: intl.formatMessage(messages.os),
-      key: 'osVersion',
-      transforms: [sortable, cellWidth(10)],
-    },
-    {
-      title: intl.formatMessage(messages.lastMatched),
-      transforms: [sortable, cellWidth(10)],
-    },
-    {
-      title: intl.formatMessage(messages.totalMatches),
-      transforms: [
-        info({
-          tooltip: intl.formatMessage(messages.totalMatchesNote),
-        }),
-        sortable,
-        cellWidth(10),
-      ],
-    },
-  ];
+
+  const fetchDetailsTable = async (shouldShowLoading) => {
+    if (shouldShowLoading) {
+      setSigDetailsTableLoading(true);
+    }
+
+    try {
+      const response = await client.query({
+        query: GET_SIGNATURE_DETAILS_TABLE,
+        fetchPolicy: 'no-cache',
+        variables: tableVars,
+      });
+
+      setSigDetailsTable(response);
+    } catch (error) {
+      console.error('Error fetching data:', error);
+    }
+
+    setSigDetailsTableLoading(false);
+  };
+
+  const fetchDetailsTableGroups = async (shouldShowLoading) => {
+    if (shouldShowLoading) {
+      setSigDetailsTableGroupsLoading(true);
+    }
+
+    let response;
+    try {
+      response = await client.query({
+        query: GET_SIGNATURE_DETAILS_TABLE_GROUPS,
+        fetchPolicy: 'no-cache',
+        variables: tableVars,
+      });
+    } catch (error) {
+      console.error('Error fetching data:', error);
+    }
+
+    setSigDetailsTableGroups(response);
+    setSigDetailsTableGroupsLoading(false);
+  };
+
+  useEffect(() => {
+    fetchDetailsTable(true);
+    fetchDetailsTableGroups(true);
+  }, [tableVars]);
 
   const page = tableVars.offset / tableVars.limit + 1;
+  const columns = sigDetailsTableColumns(isWorkspaceEnabled, intl);
 
   const filterConfigItems = (isWorkspaceEnabled) => [
     {
@@ -162,7 +187,7 @@ const SigDetailsTable = ({ ruleName, affectedCount, isEmptyAccount }) => {
             offset: 0,
           });
         },
-        items: groupsLoading
+        items: sigDetailsTableGroupsLoading
           ? [] // display an empty box whilst the groups are loading
           : isEmpty(groups)
           ? [
@@ -182,6 +207,25 @@ const SigDetailsTable = ({ ruleName, affectedCount, isEmptyAccount }) => {
             }),
       },
     },
+    {
+      label: intl.formatMessage(messages.sigMatchStatus).toLowerCase(),
+      type: 'checkbox',
+      filterValues: {
+        key: 'sig-match-status-filter',
+        onChange: (e, value) => {
+          stateSet({
+            type: 'setTableVars',
+            payload: { sigMatchStatusFilter: value },
+            offset: 0,
+          });
+        },
+        items: filterValues,
+        value: tableVars.sigMatchStatusFilter,
+        placeholder: intl.formatMessage(messages.filterBy, {
+          field: intl.formatMessage(messages.status).toLowerCase(),
+        }),
+      },
+    },
   ];
 
   const onSetPage = (e, page) =>
@@ -193,108 +237,29 @@ const SigDetailsTable = ({ ruleName, affectedCount, isEmptyAccount }) => {
   const onPerPageSelect = (e, perPage) =>
     stateSet({ type: 'setTableVars', payload: { limit: perPage, offset: 0 } });
 
-  const onSort = (e, index, direction) =>
-    stateSet({
-      type: 'setSortBy',
-      payload: { index, direction },
-      tableVars: { orderBy: orderBy({ index, direction }), offset: 0 },
-    });
-
-  const onCollapse = (e, rowKey, isOpen) => {
-    const collapseRows = [...rows];
-    const host = collapseRows[rowKey + 1].hostData;
-
-    collapseRows[rowKey] = { ...collapseRows[rowKey], isOpen };
-    collapseRows[rowKey + 1].cells = [
-      {
-        title: (
-          <CodeEditor
-            height="250px"
-            code={`${host.matches
-              .map(
-                (match, key) => `Match Source: ${match.source}
-Offset:${match.stringOffset}
-Match Data: ${match.stringData}
-Match Identifier: ${match.stringIdentifier}
-Match Scan Date: ${new Date(match.scanDate).toUTCString()}
-${match.metadata ? expandMatchMetadata(JSON.parse(match.metadata)) : ''}
-${
-  host.matches.length > 1 && key !== host.matches.length - 1
-    ? `~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-`
-    : ''
-}`
-              )
-              .join('')}`}
-            isDownloadEnabled
-            isCopyEnabled
-          />
-        ),
-      },
-    ];
-    stateSet({ type: 'setRows', payload: collapseRows });
-  };
-
-  useEffect(() => {
-    const rowBuilder = (data) =>
-      data?.flatMap((data, key) => {
-        const host = data;
-        return [
-          {
-            rowId: key,
-            isOpen: false,
-            cells: [
-              {
-                title: (
-                  <InsightsLink to={`/systems/${host.id}`}>
-                    {host.displayName}
-                  </InsightsLink>
-                ),
-              },
-              {
-                title: `${
-                  host.groups?.[0]?.name ||
-                  intl.formatMessage(messages.notApplicable)
-                }`,
-              },
-              {
-                title: host.osVersion
-                  ? `RHEL ${host.osVersion}`
-                  : intl.formatMessage(messages.dataNotAvailable),
-              },
-              {
-                title: (
-                  <DateFormat
-                    date={
-                      new Date(host.matches[host.matches.length - 1].scanDate)
-                    }
-                    type="onlyDate"
-                  />
-                ),
-              },
-              {
-                title: (
-                  <InsightsLink to={`/systems/${host.id}`}>
-                    {host.matchCount?.toLocaleString()}
-                  </InsightsLink>
-                ),
-              },
-            ],
-          },
-          {
-            parent: key * 2,
-            hostData: host,
-            fullWidth: true,
-            cells: [],
-          },
-        ];
+  const getSortParams = (columnIndex) => ({
+    sortBy: {
+      index: sortBy.index,
+      direction: sortBy.direction,
+    },
+    onSort: (e, index, direction) => {
+      stateSet({
+        type: 'setSortBy',
+        payload: { index, direction },
+        tableVars: { orderBy: orderBy({ index, direction }), offset: 0 },
       });
+    },
+    columnIndex,
+  });
 
-    stateSet({
-      type: 'setRows',
-      payload: rowBuilder(data?.rulesList[0]?.affectedHostsList),
-    });
-  }, [intl, data]);
+  const actions = useActionsConfig(
+    isWriteAccessLoading,
+    hasWriteAccess,
+    selectedMatches,
+    setSelectedMatches,
+    fetchDetailsTable,
+    fetchDetailsTableGroups
+  );
 
   useEffect(() => {
     // populates the Group name filter dropdown
@@ -308,14 +273,15 @@ ${
     };
 
     let groupsList =
-      getGroupList(groupsData?.rulesList[0]?.affectedHostsList) || [];
+      getGroupList(
+        sigDetailsTableGroups?.data?.rulesList[0]?.affectedHostsList
+      ) || [];
     groupsList = uniqWith(
       groupsList,
       (group1, group2) => group1.label === group2.label
     ); // remove duplicate group names
     stateSet({ type: 'setGroups', payload: groupsList });
-  }, [groupsData]);
-  const isWorkspaceEnabled = useWorkspaceFeatureFlag();
+  }, [sigDetailsTableGroups?.data]);
 
   const NoResultsMatch = (
     <MessageState
@@ -345,7 +311,7 @@ ${
     />
   );
 
-  const ErrorState = (
+  const ErrorState = () => (
     <MessageState
       className="pf-c-card"
       variant="large"
@@ -373,6 +339,15 @@ ${
           value: group,
         })),
       });
+    tableVars?.sigMatchStatusFilter &&
+      chips.push({
+        category: intl.formatMessage(messages.status),
+        value: 'status',
+        chips: tableVars.sigMatchStatusFilter.map((status) => ({
+          name: matchStatusFilterMap[status],
+          value: status,
+        })),
+      });
     return chips;
   };
 
@@ -380,13 +355,21 @@ ${
     deleteTitle: intl.formatMessage(messages.resetFilters),
     filters: buildFilterChips(),
     showDeleteButton:
-      tableVars?.displayName !== '' || tableVars?.hostGroupFilter !== undefined,
+      tableVars?.displayName !== '' ||
+      tableVars?.hostGroupFilter !== undefined ||
+      tableVars?.sigMatchStatusFilter !== undefined,
     onDelete: (event, itemsToRemove, isAll) => {
       if (isAll) {
         stateSet({
           type: 'setTableVars',
-          payload: { displayName: '', hostGroupFilter: undefined, offset: 0 },
+          payload: {
+            displayName: '',
+            hostGroupFilter: undefined,
+            sigMatchStatusFilter: undefined,
+            offset: 0,
+          },
         });
+        setExpandedRows([]);
       } else {
         itemsToRemove.map((item) => {
           item.value === 'name' &&
@@ -400,6 +383,17 @@ ${
               offset: 0,
             };
             stateSet({ type: 'setTableVars', payload });
+          } else if (item.value === 'status') {
+            const sigMatchStatusChips = tableVars.sigMatchStatusFilter.filter(
+              (status) => status !== item.chips[0].value
+            );
+            const payload = {
+              sigMatchStatusFilter: sigMatchStatusChips.length
+                ? sigMatchStatusChips
+                : undefined,
+              offset: 0,
+            };
+            stateSet({ type: 'setTableVars', payload });
           }
         });
       }
@@ -409,8 +403,13 @@ ${
   return (
     <React.Fragment>
       <PrimaryToolbar
+        actionsConfig={{
+          dropdownProps: { popperProps: { position: 'left' } },
+          actions,
+        }}
         pagination={{
-          itemCount: data?.rulesList[0]?.affectedHosts?.totalCount || 0,
+          itemCount:
+            sigDetailsTable?.data?.rulesList[0]?.affectedHosts?.totalCount || 0,
           page,
           perPage: tableVars.limit,
           onSetPage(e, page) {
@@ -424,44 +423,51 @@ ${
         filterConfig={{ items: filterConfigItems(isWorkspaceEnabled) }}
         activeFiltersConfig={activeFiltersConfig}
       />
-      {loading ? (
+      {sigDetailsTableLoading ? (
         <SkeletonTable
           rows={tableVars.limit}
-          columns={columns(isWorkspaceEnabled).map((column) => column.title)}
+          columns={columns.map((column) => column.title)}
         />
       ) : (
         <Table
           className="sigTable"
           aria-label="Signature Details table"
-          rows={rows}
-          cells={columns(isWorkspaceEnabled)}
-          onCollapse={onCollapse}
-          onSort={onSort}
-          sortBy={
-            data?.rulesList[0]?.affectedHosts?.totalCount > 0
-              ? sortBy
-              : undefined
-          }
           isStickyHeader
         >
-          <TableHeader />
-          <TableBody />
+          <DetailsTableHeader columns={columns} getSortParams={getSortParams} />
+          <SigDetailsTableBody
+            columns={columns}
+            data={sigDetailsTable.data?.rulesList[0]?.affectedHostsList}
+            expandedRows={expandedRows}
+            fetchDetailsTable={fetchDetailsTable}
+            hasWriteAccess={hasWriteAccess}
+            isWriteAccessLoading={isWriteAccessLoading}
+            ruleName={tableVars.ruleName}
+            setExpandedRows={setExpandedRows}
+            setSelectedMatches={setSelectedMatches}
+            sigMatchStatusFilter={tableVars.sigMatchStatusFilter}
+          />
         </Table>
       )}
 
-      {!error && !loading && affectedCount !== 0 ? (
-        data?.rulesList[0]?.affectedHosts?.totalCount === 0 ? (
+      {!sigDetailsTable?.error &&
+      !sigDetailsTableLoading &&
+      affectedCount !== 0 ? (
+        sigDetailsTable?.data?.rulesList[0]?.affectedHosts?.totalCount === 0 ? (
           NoResultsMatch
         ) : (
           <React.Fragment />
         )
       ) : (
-        !loading && (isEmptyAccount ? EmptyAccountState : NoAffectedHosts)
+        !sigDetailsTableLoading &&
+        (isEmptyAccount ? EmptyAccountState : NoAffectedHosts)
       )}
-      {error && <ErrorState />}
+      {sigDetailsTable?.error && <ErrorState />}
       <TableToolbar isFooter>
         <Pagination
-          itemCount={data?.rulesList[0]?.affectedHosts?.totalCount || 0}
+          itemCount={
+            sigDetailsTable?.data?.rulesList[0]?.affectedHosts?.totalCount || 0
+          }
           widgetId="pagination-options-menu-bottom"
           perPage={tableVars.limit}
           page={page}
