+++ src/Components/SigDetailsTable/__tests__/SigDetailsTable.tests.js
@@ -6,7 +6,7 @@ import { IntlProvider } from 'react-intl';
 import { useApolloClient, useMutation } from '@apollo/client';
 import '@testing-library/jest-dom';
 
-import NewSigDetailsTable from '../NewSigDetailsTable';
+import SigDetailsTable from '../SigDetailsTable';
 import {
   mockSigDetailsTableData,
   mockSigHitsList,
@@ -46,7 +46,7 @@ describe('SigDetailsTable', () => {
     render(
       <IntlProvider locale="en">
         <Router>
-          <NewSigDetailsTable
+          <SigDetailsTable
             ruleName="bingo bango bongo"
             affectedCount={1}
             isEmptyAccount={false}
@@ -66,7 +66,7 @@ describe('SigDetailsTable', () => {
     render(
       <IntlProvider locale="en">
         <Router>
-          <NewSigDetailsTable
+          <SigDetailsTable
             ruleName="bingo bango bongo"
             affectedCount={1}
             isEmptyAccount={false}
@@ -110,7 +110,7 @@ describe('SigDetailsTable', () => {
     render(
       <IntlProvider locale="en">
         <Router>
-          <NewSigDetailsTable
+          <SigDetailsTable
             ruleName="bingo bango bongo"
             affectedCount={1}
             isEmptyAccount={false}
@@ -180,7 +180,7 @@ describe('SigDetailsTable', () => {
     render(
       <IntlProvider locale="en">
         <Router>
-          <NewSigDetailsTable
+          <SigDetailsTable
             ruleName="bingo bango bongo"
             affectedCount={1}
             isEmptyAccount={false}
@@ -214,7 +214,7 @@ describe('SigDetailsTable', () => {
     render(
       <IntlProvider locale="en">
         <Router>
-          <NewSigDetailsTable
+          <SigDetailsTable
             ruleName="bingo bango bongo"
             affectedCount={1}
             isEmptyAccount={false}
