--- src/Components/SysDetailsTable/NewSysDetailsTable.js
@@ -1,345 +1,0 @@
-import React, { useEffect, useReducer, useState } from 'react';
-import propTypes from 'prop-types';
-import { useIntl } from 'react-intl';
-import { useApolloClient } from '@apollo/client';
-import { Pagination, PaginationVariant } from '@patternfly/react-core';
-import { Table } from '@patternfly/react-table';
-import { SortByDirection } from '@patternfly/react-table';
-import { CheckCircleIcon, SearchIcon } from '@patternfly/react-icons';
-import { SkeletonTable } from '@patternfly/react-component-groups';
-import TableToolbar from '@redhat-cloud-services/frontend-components/TableToolbar';
-import { PrimaryToolbar } from '@redhat-cloud-services/frontend-components/PrimaryToolbar';
-import { GET_SYSTEMS_DETAILS_TABLE_PAGE } from '../../operations/queries';
-import MessageState from '../MessageState/MessageState';
-import messages from '../../Messages';
-import { matchStatusFilterMap } from '../helpers';
-import { sysDetailsTableColumns } from './Columns';
-import { useActionsConfig } from '../../Utilities/Hooks';
-import { filterValues } from '../constants';
-import { usePermissions } from '@redhat-cloud-services/frontend-components-utilities/RBACHook';
-import { RBACPermissions } from '../Common';
-import { DetailsTableHeader } from '../TableComponents/DetailsTableHeader';
-import { SysDetailsTableBody } from './SysDetailsTableBody';
-
-const sortIndices = {
-  1: 'NAME',
-  2: 'LAST_STATUS',
-  3: 'LAST_MATCH_DATE',
-  4: 'NEW_MATCH_COUNT',
-  5: 'MATCH_COUNT',
-};
-
-const orderBy = ({ index, direction }) =>
-  `${sortIndices[index]}_${direction === SortByDirection.asc ? 'ASC' : 'DESC'}`;
-
-const tableReducer = (state, action) => {
-  switch (action.type) {
-    case 'setTableVars':
-      return {
-        ...state,
-        tableVars:
-          action.payload.sysMatchStatusFilter?.length === 0
-            ? {
-                ...state.tableVars,
-                sysMatchStatusFilter: undefined,
-              }
-            : {
-                ...state.tableVars,
-                ...action.payload,
-              },
-      };
-    case 'setSortBy':
-      return {
-        ...state,
-        sortBy: action.payload,
-        tableVars: { ...state.tableVars, ...action.tableVars },
-      };
-    case 'setRows':
-      return { ...state, rows: action.payload };
-  }
-
-  return state;
-};
-
-const NewSysDetailsTable = ({ systemId }) => {
-  const intl = useIntl();
-  const client = useApolloClient();
-  const { hasAccess: hasWriteAccess, isLoading: isWriteAccessLoading } =
-    usePermissions('malware-detection', RBACPermissions.write);
-  const [selectedMatches, setSelectedMatches] = useState([]);
-  const [sysDetailsTableLoading, setSysDetailsTableLoading] = useState(false);
-  const [sysDetailsTable, setSysDetailsTable] = useState({});
-  const [expandedRows, setExpandedRows] = useState([]);
-
-  const initialState = {
-    tableVars: {
-      limit: 10,
-      offset: 0,
-      orderBy: 'LAST_MATCH_DATE_ASC',
-      ruleName: '',
-      sysMatchStatusFilter: undefined,
-      systemId,
-    },
-    sortBy: {
-      index: 2,
-      direction: SortByDirection.asc,
-    },
-    rows: [],
-  };
-
-  const [{ tableVars, sortBy }, stateSet] = useReducer(tableReducer, {
-    ...initialState,
-  });
-
-  const fetchDetailsTable = async (shouldShowLoading) => {
-    if (shouldShowLoading) {
-      setSysDetailsTableLoading(true);
-    }
-
-    try {
-      const response = await client.query({
-        query: GET_SYSTEMS_DETAILS_TABLE_PAGE,
-        fetchPolicy: 'no-cache',
-        variables: tableVars,
-      });
-
-      setSysDetailsTable(response);
-    } catch (error) {
-      console.error('Error fetching data:', error);
-    }
-
-    setSysDetailsTableLoading(false);
-  };
-
-  useEffect(() => {
-    fetchDetailsTable(true);
-  }, [tableVars]);
-
-  const page = tableVars.offset / tableVars.limit + 1;
-  const columns = sysDetailsTableColumns(intl);
-
-  const filterConfigItems = [
-    {
-      label: intl.formatMessage(messages.sig).toLowerCase(),
-      type: 'text',
-      filterValues: {
-        key: 'text-filter',
-        onChange: (e, value) =>
-          stateSet({
-            type: 'setTableVars',
-            payload: { ruleName: value, offset: 0 },
-          }),
-        value: tableVars.ruleName,
-        placeholder: intl.formatMessage(messages.filterBy, {
-          field: intl.formatMessage(messages.sig).toLowerCase(),
-        }),
-      },
-    },
-    {
-      label: intl.formatMessage(messages.sysMatchStatus).toLowerCase(),
-      type: 'checkbox',
-      filterValues: {
-        key: 'sys-match-status-filter',
-        onChange: (e, value) => {
-          stateSet({
-            type: 'setTableVars',
-            payload: { sysMatchStatusFilter: value },
-            offset: 0,
-          });
-        },
-        items: filterValues,
-        value: tableVars.sysMatchStatusFilter,
-        placeholder: intl.formatMessage(messages.filterBy, {
-          field: intl.formatMessage(messages.status).toLowerCase(),
-        }),
-      },
-    },
-  ];
-
-  const onSetPage = (e, page) =>
-    stateSet({
-      type: 'setTableVars',
-      payload: { offset: page * tableVars.limit - tableVars.limit },
-    });
-
-  const onPerPageSelect = (e, perPage) =>
-    stateSet({ type: 'setTableVars', payload: { limit: perPage, offset: 0 } });
-
-  const getSortParams = (columnIndex) => ({
-    sortBy: {
-      index: sortBy.index,
-      direction: sortBy.direction,
-    },
-    onSort: (e, index, direction) => {
-      stateSet({
-        type: 'setSortBy',
-        payload: { index, direction },
-        tableVars: { orderBy: orderBy({ index, direction }), offset: 0 },
-      });
-    },
-    columnIndex,
-  });
-
-  const actions = useActionsConfig(
-    isWriteAccessLoading,
-    hasWriteAccess,
-    selectedMatches,
-    setSelectedMatches,
-    fetchDetailsTable
-  );
-
-  const buildFilterChips = () => {
-    const chips = [];
-    tableVars?.ruleName &&
-      chips.push({
-        category: intl.formatMessage(messages.signature),
-        value: 'signature',
-        chips: [{ name: tableVars.ruleName, value: tableVars.ruleName }],
-      });
-    tableVars?.sysMatchStatusFilter &&
-      chips.push({
-        category: intl.formatMessage(messages.status),
-        value: 'status',
-        chips: tableVars.sysMatchStatusFilter.map((status) => ({
-          name: matchStatusFilterMap[status],
-          value: status,
-        })),
-      });
-
-    return chips;
-  };
-
-  const activeFiltersConfig = {
-    deleteTitle: intl.formatMessage(messages.resetFilters),
-    filters: buildFilterChips(),
-    showDeleteButton:
-      tableVars?.ruleName !== '' ||
-      tableVars?.sysMatchStatusFilter !== undefined,
-    onDelete: (event, itemsToRemove, isAll) => {
-      if (isAll) {
-        stateSet({
-          type: 'setTableVars',
-          payload: { ruleName: '', sysMatchStatusFilter: undefined, offset: 0 },
-        });
-        setExpandedRows([]);
-      } else {
-        itemsToRemove.map((item) => {
-          if (item.value === 'signature') {
-            stateSet({ type: 'setTableVars', payload: { ruleName: '' } });
-          } else if (item.value === 'status') {
-            const sysMatchStatusChips = tableVars.sysMatchStatusFilter.filter(
-              (status) => status !== item.chips[0].value
-            );
-            const payload = {
-              sysMatchStatusFilter: sysMatchStatusChips.length
-                ? sysMatchStatusChips
-                : undefined,
-              offset: 0,
-            };
-            stateSet({ type: 'setTableVars', payload });
-          }
-        });
-      }
-    },
-  };
-
-  return (
-    <React.Fragment>
-      <PrimaryToolbar
-        actionsConfig={{
-          dropdownProps: { popperProps: { position: 'left' } },
-          actions,
-        }}
-        pagination={{
-          itemCount:
-            sysDetailsTable?.data?.host?.affectedRules?.totalCount || 0,
-          page,
-          perPage: tableVars.limit,
-          onSetPage(e, page) {
-            onSetPage(e, page);
-          },
-          onPerPageSelect(e, perPage) {
-            onPerPageSelect(e, perPage);
-          },
-          isCompact: true,
-        }}
-        filterConfig={{ items: filterConfigItems }}
-        activeFiltersConfig={activeFiltersConfig}
-      />
-      {!sysDetailsTableLoading && (
-        <Table
-          className="sigTable"
-          aria-label="Signature Details table"
-          isStickyHeader
-        >
-          <DetailsTableHeader columns={columns} getSortParams={getSortParams} />
-          <SysDetailsTableBody
-            columns={columns}
-            data={sysDetailsTable.data?.host}
-            expandedRows={expandedRows}
-            fetchDetailsTable={fetchDetailsTable}
-            hasWriteAccess={hasWriteAccess}
-            isWriteAccessLoading={isWriteAccessLoading}
-            setExpandedRows={setExpandedRows}
-            setSelectedMatches={setSelectedMatches}
-            sysMatchStatusFilter={tableVars.sysMatchStatusFilter}
-          />
-        </Table>
-      )}
-
-      {sysDetailsTableLoading ? (
-        <SkeletonTable
-          rows={tableVars.limit}
-          columns={columns.map((column) => column.title)}
-        />
-      ) : !sysDetailsTable?.error ? (
-        sysDetailsTable?.data?.host?.lastMatchDate ? (
-          sysDetailsTable?.data?.host?.affectedRules?.totalCount === 0 ? (
-            <MessageState
-              className="pf-c-card"
-              icon={SearchIcon}
-              variant="large"
-              title={intl.formatMessage(messages.noResults)}
-              text={intl.formatMessage(messages.noResultsMatch)}
-            />
-          ) : (
-            <React.Fragment />
-          )
-        ) : (
-          <MessageState
-            className="pf-c-card"
-            variant="large"
-            icon={CheckCircleIcon}
-            iconClass="ins-l-success-color"
-            title={intl.formatMessage(messages.notMatched)}
-            text={intl.formatMessage(messages.notMatchedBody)}
-          />
-        )
-      ) : (
-        <MessageState
-          className="pf-c-card"
-          variant="large"
-          title="Error"
-          text="error"
-        />
-      )}
-      <TableToolbar isFooter>
-        <Pagination
-          itemCount={
-            sysDetailsTable?.data?.host?.affectedRules?.totalCount || 0
-          }
-          widgetId="pagination-options-menu-bottom"
-          perPage={tableVars.limit}
-          page={page}
-          variant={PaginationVariant.bottom}
-          onSetPage={onSetPage}
-          onPerPageSelect={onPerPageSelect}
-        />
-      </TableToolbar>
-    </React.Fragment>
-  );
-};
-
-NewSysDetailsTable.propTypes = { systemId: propTypes.string };
-
-export default NewSysDetailsTable;
