+++ src/Components/SysDetailsTable/SysDetailsTable.js
@@ -1,49 +1,53 @@
-import React, { useEffect, useReducer } from 'react';
+import React, { useEffect, useReducer, useState } from 'react';
 import propTypes from 'prop-types';
 import { useIntl } from 'react-intl';
-import { useQuery } from '@apollo/client';
-import {
-  Pagination,
-  PaginationVariant,
-} from '@patternfly/react-core/dist/esm/components/Pagination';
-import {
-  Table,
-  TableBody,
-  TableHeader,
-} from '@patternfly/react-table/deprecated';
-import {
-  SortByDirection,
-  cellWidth,
-  expandable,
-  sortable,
-  info,
-} from '@patternfly/react-table';
+import { useApolloClient } from '@apollo/client';
+import { Pagination, PaginationVariant } from '@patternfly/react-core';
+import { Table } from '@patternfly/react-table';
+import { SortByDirection } from '@patternfly/react-table';
 import { CheckCircleIcon, SearchIcon } from '@patternfly/react-icons';
-import SkeletonTable from '@patternfly/react-component-groups/dist/esm/SkeletonTable';
-import InsightsLink from '@redhat-cloud-services/frontend-components/InsightsLink';
+import { SkeletonTable } from '@patternfly/react-component-groups';
 import TableToolbar from '@redhat-cloud-services/frontend-components/TableToolbar';
 import { PrimaryToolbar } from '@redhat-cloud-services/frontend-components/PrimaryToolbar';
-import CodeEditor from '../CodeEditor/CodeEditor';
-import { DateFormat } from '@redhat-cloud-services/frontend-components/DateFormat';
-import { expandMatchMetadata } from '../Common';
 import { GET_SYSTEMS_DETAILS_TABLE_PAGE } from '../../operations/queries';
 import MessageState from '../MessageState/MessageState';
-import StatusLabel from '../StatusLabel/StatusLabel';
 import messages from '../../Messages';
+import { matchStatusFilterMap } from '../helpers';
+import { sysDetailsTableColumns } from './Columns';
+import { useActionsConfig } from '../../Utilities/Hooks';
+import { filterValues } from '../constants';
+import { usePermissions } from '@redhat-cloud-services/frontend-components-utilities/RBACHook';
+import { RBACPermissions } from '../Common';
+import { DetailsTableHeader } from '../TableComponents/DetailsTableHeader';
+import { SysDetailsTableBody } from './SysDetailsTableBody';
 
 const sortIndices = {
   1: 'NAME',
   2: 'LAST_STATUS',
   3: 'LAST_MATCH_DATE',
-  4: 'MATCH_COUNT',
+  4: 'NEW_MATCH_COUNT',
+  5: 'MATCH_COUNT',
 };
+
 const orderBy = ({ index, direction }) =>
   `${sortIndices[index]}_${direction === SortByDirection.asc ? 'ASC' : 'DESC'}`;
 
 const tableReducer = (state, action) => {
   switch (action.type) {
     case 'setTableVars':
-      return { ...state, tableVars: { ...state.tableVars, ...action.payload } };
+      return {
+        ...state,
+        tableVars:
+          action.payload.sysMatchStatusFilter?.length === 0
+            ? {
+                ...state.tableVars,
+                sysMatchStatusFilter: undefined,
+              }
+            : {
+                ...state.tableVars,
+                ...action.payload,
+              },
+      };
     case 'setSortBy':
       return {
         ...state,
@@ -59,12 +63,21 @@ const tableReducer = (state, action) => {
 
 const SysDetailsTable = ({ systemId }) => {
   const intl = useIntl();
+  const client = useApolloClient();
+  const { hasAccess: hasWriteAccess, isLoading: isWriteAccessLoading } =
+    usePermissions('malware-detection', RBACPermissions.write);
+  const [selectedMatches, setSelectedMatches] = useState([]);
+  const [sysDetailsTableLoading, setSysDetailsTableLoading] = useState(false);
+  const [sysDetailsTable, setSysDetailsTable] = useState({});
+  const [expandedRows, setExpandedRows] = useState([]);
+
   const initialState = {
     tableVars: {
       limit: 10,
       offset: 0,
       orderBy: 'LAST_MATCH_DATE_ASC',
       ruleName: '',
+      sysMatchStatusFilter: undefined,
       systemId,
     },
     sortBy: {
@@ -73,39 +86,37 @@ const SysDetailsTable = ({ systemId }) => {
     },
     rows: [],
   };
-  const [{ tableVars, sortBy, rows }, stateSet] = useReducer(tableReducer, {
+
+  const [{ tableVars, sortBy }, stateSet] = useReducer(tableReducer, {
     ...initialState,
   });
-  const { data, loading, error } = useQuery(GET_SYSTEMS_DETAILS_TABLE_PAGE, {
-    variables: tableVars,
-  });
-  const columns = [
-    {
-      title: intl.formatMessage(messages.sigName),
-      cellFormatters: [expandable],
-      transforms: [sortable, cellWidth(55)],
-    },
-    {
-      title: intl.formatMessage(messages.lastStatus),
-      transforms: [cellWidth(10), sortable],
-    },
-    {
-      title: intl.formatMessage(messages.matched),
-      transforms: [sortable, cellWidth(10)],
-    },
-    {
-      title: intl.formatMessage(messages.totalMatches),
-      transforms: [
-        cellWidth(15),
-        info({
-          tooltip: intl.formatMessage(messages.totalMatchesNote),
-        }),
-        sortable,
-      ],
-    },
-  ];
+
+  const fetchDetailsTable = async (shouldShowLoading) => {
+    if (shouldShowLoading) {
+      setSysDetailsTableLoading(true);
+    }
+
+    try {
+      const response = await client.query({
+        query: GET_SYSTEMS_DETAILS_TABLE_PAGE,
+        fetchPolicy: 'no-cache',
+        variables: tableVars,
+      });
+
+      setSysDetailsTable(response);
+    } catch (error) {
+      console.error('Error fetching data:', error);
+    }
+
+    setSysDetailsTableLoading(false);
+  };
+
+  useEffect(() => {
+    fetchDetailsTable(true);
+  }, [tableVars]);
 
   const page = tableVars.offset / tableVars.limit + 1;
+  const columns = sysDetailsTableColumns(intl);
 
   const filterConfigItems = [
     {
@@ -124,6 +135,25 @@ const SysDetailsTable = ({ systemId }) => {
         }),
       },
     },
+    {
+      label: intl.formatMessage(messages.sysMatchStatus).toLowerCase(),
+      type: 'checkbox',
+      filterValues: {
+        key: 'sys-match-status-filter',
+        onChange: (e, value) => {
+          stateSet({
+            type: 'setTableVars',
+            payload: { sysMatchStatusFilter: value },
+            offset: 0,
+          });
+        },
+        items: filterValues,
+        value: tableVars.sysMatchStatusFilter,
+        placeholder: intl.formatMessage(messages.filterBy, {
+          field: intl.formatMessage(messages.status).toLowerCase(),
+        }),
+      },
+    },
   ];
 
   const onSetPage = (e, page) =>
@@ -135,128 +165,94 @@ const SysDetailsTable = ({ systemId }) => {
   const onPerPageSelect = (e, perPage) =>
     stateSet({ type: 'setTableVars', payload: { limit: perPage, offset: 0 } });
 
-  const onSort = (e, index, direction) =>
-    stateSet({
-      type: 'setSortBy',
-      payload: { index, direction },
-      tableVars: { orderBy: orderBy({ index, direction }), offset: 0 },
-    });
+  const getSortParams = (columnIndex) => ({
+    sortBy: {
+      index: sortBy.index,
+      direction: sortBy.direction,
+    },
+    onSort: (e, index, direction) => {
+      stateSet({
+        type: 'setSortBy',
+        payload: { index, direction },
+        tableVars: { orderBy: orderBy({ index, direction }), offset: 0 },
+      });
+    },
+    columnIndex,
+  });
 
-  const onCollapse = (e, rowKey, isOpen) => {
-    const collapseRows = [...rows];
-    const host = collapseRows[rowKey + 1].hostData;
-
-    collapseRows[rowKey] = { ...collapseRows[rowKey], isOpen };
-    collapseRows[rowKey + 1].cells = [
-      {
-        title: (
-          <CodeEditor
-            height="250px"
-            code={`${host.matches
-              .map(
-                (match, key) => `Match Source: ${match.source}
-Offset:${match.stringOffset}
-Match Data: ${match.stringData}
-Match Identifier: ${match.stringIdentifier}
-Match Scan Date: ${new Date(match.scanDate).toUTCString()}
-${
-  match.metadata
-    ? expandMatchMetadata(JSON.parse(match.metadata.replaceAll('\n', '')))
-    : ''
-}
-${
-  host.matches.length > 1 && key !== host.matches.length - 1
-    ? `~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-`
-    : ''
-}`
-              )
-              .join('')}`}
-            isDownloadEnabled
-            isCopyEnabled
-          />
-        ),
-      },
-    ];
-    stateSet({ type: 'setRows', payload: collapseRows });
-  };
-
-  useEffect(() => {
-    const rowBuilder = (data) =>
-      data?.host.affectedRulesList.flatMap((data, key) => [
-        {
-          rowId: key,
-          isOpen: false,
-          cells: [
-            {
-              title: (
-                <InsightsLink to={`/signatures/${data.name}`}>
-                  {data.name}
-                </InsightsLink>
-              ),
-            },
-            {
-              title: (
-                <StatusLabel
-                  isDisabled={data.isDisabled}
-                  hasMatch={data.lastStatus}
-                  displayMatch
-                />
-              ),
-            },
-            {
-              title: (
-                <DateFormat
-                  date={new Date(data.lastMatchDate)}
-                  type="onlyDate"
-                />
-              ),
-            },
-            {
-              title: (
-                <InsightsLink to={`/signatures/${data.name}`}>
-                  {data.matchCount?.toLocaleString()}
-                </InsightsLink>
-              ),
-            },
-          ],
-        },
-        {
-          parent: key * 2,
-          hostData: data,
-          fullWidth: true,
-          cells: [],
-        },
-      ]);
-
-    !loading && stateSet({ type: 'setRows', payload: rowBuilder(data) });
-  }, [data, loading]);
+  const actions = useActionsConfig(
+    isWriteAccessLoading,
+    hasWriteAccess,
+    selectedMatches,
+    setSelectedMatches,
+    fetchDetailsTable
+  );
 
   const buildFilterChips = () => {
-    return tableVars?.ruleName
-      ? [
-          {
-            category: intl.formatMessage(messages.signature),
-            value: 'signature',
-            chips: [{ name: tableVars.ruleName, value: tableVars.ruleName }],
-          },
-        ]
-      : [];
+    const chips = [];
+    tableVars?.ruleName &&
+      chips.push({
+        category: intl.formatMessage(messages.signature),
+        value: 'signature',
+        chips: [{ name: tableVars.ruleName, value: tableVars.ruleName }],
+      });
+    tableVars?.sysMatchStatusFilter &&
+      chips.push({
+        category: intl.formatMessage(messages.status),
+        value: 'status',
+        chips: tableVars.sysMatchStatusFilter.map((status) => ({
+          name: matchStatusFilterMap[status],
+          value: status,
+        })),
+      });
+
+    return chips;
   };
 
   const activeFiltersConfig = {
     deleteTitle: intl.formatMessage(messages.resetFilters),
     filters: buildFilterChips(),
-    showDeleteButton: tableVars?.ruleName !== '',
-    onDelete: () =>
-      stateSet({ type: 'setTableVars', payload: { ruleName: '', offset: 0 } }),
+    showDeleteButton:
+      tableVars?.ruleName !== '' ||
+      tableVars?.sysMatchStatusFilter !== undefined,
+    onDelete: (event, itemsToRemove, isAll) => {
+      if (isAll) {
+        stateSet({
+          type: 'setTableVars',
+          payload: { ruleName: '', sysMatchStatusFilter: undefined, offset: 0 },
+        });
+        setExpandedRows([]);
+      } else {
+        itemsToRemove.map((item) => {
+          if (item.value === 'signature') {
+            stateSet({ type: 'setTableVars', payload: { ruleName: '' } });
+          } else if (item.value === 'status') {
+            const sysMatchStatusChips = tableVars.sysMatchStatusFilter.filter(
+              (status) => status !== item.chips[0].value
+            );
+            const payload = {
+              sysMatchStatusFilter: sysMatchStatusChips.length
+                ? sysMatchStatusChips
+                : undefined,
+              offset: 0,
+            };
+            stateSet({ type: 'setTableVars', payload });
+          }
+        });
+      }
+    },
   };
 
   return (
     <React.Fragment>
       <PrimaryToolbar
+        actionsConfig={{
+          dropdownProps: { popperProps: { position: 'left' } },
+          actions,
+        }}
         pagination={{
-          itemCount: data?.host?.affectedRules?.totalCount || 0,
+          itemCount:
+            sysDetailsTable?.data?.host?.affectedRules?.totalCount || 0,
           page,
           perPage: tableVars.limit,
           onSetPage(e, page) {
@@ -270,32 +266,35 @@ ${
         filterConfig={{ items: filterConfigItems }}
         activeFiltersConfig={activeFiltersConfig}
       />
-      {!loading && (
+      {!sysDetailsTableLoading && (
         <Table
           className="sigTable"
           aria-label="Signature Details table"
-          rows={rows}
-          cells={columns}
-          onCollapse={onCollapse}
-          onSort={onSort}
-          sortBy={
-            data?.host?.affectedRules?.totalCount > 0 ? sortBy : undefined
-          }
           isStickyHeader
         >
-          <TableHeader />
-          <TableBody />
+          <DetailsTableHeader columns={columns} getSortParams={getSortParams} />
+          <SysDetailsTableBody
+            columns={columns}
+            data={sysDetailsTable.data?.host}
+            expandedRows={expandedRows}
+            fetchDetailsTable={fetchDetailsTable}
+            hasWriteAccess={hasWriteAccess}
+            isWriteAccessLoading={isWriteAccessLoading}
+            setExpandedRows={setExpandedRows}
+            setSelectedMatches={setSelectedMatches}
+            sysMatchStatusFilter={tableVars.sysMatchStatusFilter}
+          />
         </Table>
       )}
 
-      {loading ? (
+      {sysDetailsTableLoading ? (
         <SkeletonTable
           rows={tableVars.limit}
           columns={columns.map((column) => column.title)}
         />
-      ) : !error ? (
-        data?.host?.lastMatchDate ? (
-          data?.host?.affectedRules?.totalCount === 0 ? (
+      ) : !sysDetailsTable?.error ? (
+        sysDetailsTable?.data?.host?.lastMatchDate ? (
+          sysDetailsTable?.data?.host?.affectedRules?.totalCount === 0 ? (
             <MessageState
               className="pf-c-card"
               icon={SearchIcon}
@@ -326,7 +325,9 @@ ${
       )}
       <TableToolbar isFooter>
         <Pagination
-          itemCount={data?.host?.affectedRules?.totalCount || 0}
+          itemCount={
+            sysDetailsTable?.data?.host?.affectedRules?.totalCount || 0
+          }
           widgetId="pagination-options-menu-bottom"
           perPage={tableVars.limit}
           page={page}
