+++ src/Components/SysDetailsTable/__tests__/SysDetailsTable.tests.js
@@ -6,7 +6,7 @@ import { IntlProvider } from 'react-intl';
 import { useApolloClient, useMutation } from '@apollo/client';
 import '@testing-library/jest-dom';
 
-import NewSysDetailsTable from '../NewSysDetailsTable';
+import SysDetailsTable from '../SysDetailsTable';
 import {
   mockSysHitsList,
   mockSysDetailsTableData,
@@ -41,7 +41,7 @@ describe('SysDetailsTable', () => {
     render(
       <IntlProvider locale="en">
         <Router>
-          <NewSysDetailsTable systemId={1234} />
+          <SysDetailsTable systemId={1234} />
         </Router>
       </IntlProvider>
     );
@@ -57,7 +57,7 @@ describe('SysDetailsTable', () => {
     render(
       <IntlProvider locale="en">
         <Router>
-          <NewSysDetailsTable systemId={1234} />
+          <SysDetailsTable systemId={1234} />
         </Router>
       </IntlProvider>
     );
@@ -115,7 +115,7 @@ describe('SysDetailsTable', () => {
     render(
       <IntlProvider locale="en">
         <Router>
-          <NewSysDetailsTable systemId={1234} />
+          <SysDetailsTable systemId={1234} />
         </Router>
       </IntlProvider>
     );
@@ -161,7 +161,7 @@ describe('SysDetailsTable', () => {
     render(
       <IntlProvider locale="en">
         <Router>
-          <NewSysDetailsTable
+          <SysDetailsTable
             ruleName="bingo bango bongo"
             affectedCount={1}
             isEmptyAccount={false}
@@ -195,7 +195,7 @@ describe('SysDetailsTable', () => {
     render(
       <IntlProvider locale="en">
         <Router>
-          <NewSysDetailsTable systemId={1234} />
+          <SysDetailsTable systemId={1234} />
         </Router>
       </IntlProvider>
     );
