+++ src/Routes/Signatures/Details.js
@@ -33,15 +33,12 @@ import {
 } from '../../operations/mutations';
 import Loading from '../../Components/Loading/Loading';
 import SigDetailsTable from '../../Components/SigDetailsTable/SigDetailsTable';
-import NewSigDetailsTable from '../../Components/SigDetailsTable/NewSigDetailsTable';
 import { RBACPermissions } from '../../Components/Common';
 import messages from '../../Messages';
 import MessageState from '../../Components/MessageState/MessageState';
 import './Details.scss';
-import { useNewMatchesFeatureFlag } from '../../Utilities/Hooks';
 
 const Details = () => {
-  const areNewMatchesEnabled = useNewMatchesFeatureFlag();
   const intl = useIntl();
   const dispatch = useDispatch();
   const notification = (data) => dispatch(addNotification(data));
@@ -363,19 +360,11 @@ const Details = () => {
             <Title className="ins-l-tableBlockHeader" headingLevel="h2">
               {intl.formatMessage(messages.affectedHosts)}
             </Title>
-            {areNewMatchesEnabled ? (
-              <NewSigDetailsTable
-                ruleName={sigName}
-                isEmptyAccount={data?.hosts?.totalCount === 0}
-                affectedCount={sigDetailsData?.affectedHosts?.totalCount}
-              />
-            ) : (
-              <SigDetailsTable
-                ruleName={sigName}
-                isEmptyAccount={data?.hosts?.totalCount === 0}
-                affectedCount={sigDetailsData?.affectedHosts?.totalCount}
-              />
-            )}
+            <SigDetailsTable
+              ruleName={sigName}
+              isEmptyAccount={data?.hosts?.totalCount === 0}
+              affectedCount={sigDetailsData?.affectedHosts?.totalCount}
+            />
           </React.Fragment>
         )}
       </Main>
