+++ src/Routes/Systems/Details.js
@@ -15,14 +15,11 @@ import { GET_SYSTEMS_DETAILS_PAGE } from '../../operations/queries';
 import Loading from '../../Components/Loading/Loading';
 import Breadcrumb from '../../Components/Breadcrumb/Breadcrumb';
 import SysDetailsTable from '../../Components/SysDetailsTable/SysDetailsTable';
-import NewSysDetailsTable from '../../Components/SysDetailsTable/NewSysDetailsTable';
 import messages from '../../Messages';
 import { totalMatchesTitle } from '../../Components/Common';
 import './Details.scss';
-import { useNewMatchesFeatureFlag } from '../../Utilities/Hooks';
 
 const Details = () => {
-  const areNewMatchesEnabled = useNewMatchesFeatureFlag();
   const intl = useIntl();
   const { id: sysId } = useParams();
   const { data, loading } = useQuery(GET_SYSTEMS_DETAILS_PAGE, {
@@ -113,8 +110,6 @@ const Details = () => {
         </Title>
         {!sysData?.id ? (
           <Loading />
-        ) : areNewMatchesEnabled ? (
-          <NewSysDetailsTable systemId={sysData?.id} />
         ) : (
           <SysDetailsTable systemId={sysData?.id} />
         )}
