+++ src/Utilities/Hooks.js
@@ -11,9 +11,6 @@ export const useFeatureFlag = (flag) => {
 export const useWorkspaceFeatureFlag = () =>
   useFeatureFlag('platform.rbac.groups-to-workspaces-rename');
 
-export const useNewMatchesFeatureFlag = () =>
-  useFeatureFlag('insights.malware.matches');
-
 export const useActionsConfig = (
   isWriteAccessLoading,
   hasWriteAccess,
